ocr_test/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pytesseract
Pillow
